[minor] refine mnist and max cut example
