+++ examples/qaoa/max_cut_paramshift.py
